% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Android_gpl}
\alias{Android_gpl}
\title{Palette Android_gpl}
\format{A vector of length 15}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Android_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
