% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cool_gpl}
\alias{Cool_gpl}
\title{Palette Cool_gpl}
\format{A vector of length 8}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Cool_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
