% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ega_gpl}
\alias{Ega_gpl}
\title{Palette Ega_gpl}
\format{A vector of length 240}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Ega_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
