% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gold_gpl}
\alias{Gold_gpl}
\title{Palette Gold_gpl}
\format{A vector of length 102}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Gold_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
