% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Greens_gpl}
\alias{Greens_gpl}
\title{Palette Greens_gpl}
\format{A vector of length 256}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Greens_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
