% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hilite_gpl}
\alias{Hilite_gpl}
\title{Palette Hilite_gpl}
\format{A vector of length 170}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Hilite_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
