% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inkscape_gpl}
\alias{Inkscape_gpl}
\title{Palette Inkscape_gpl}
\format{A vector of length 431}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Inkscape_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
