% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lights_gpl}
\alias{Lights_gpl}
\title{Palette Lights_gpl}
\format{A vector of length 25}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Lights_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
