% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MATLAB_gpl}
\alias{MATLAB_gpl}
\title{Palette MATLAB_gpl}
\format{A vector of length 73}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
MATLAB_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
