% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Paintjet_gpl}
\alias{Paintjet_gpl}
\title{Palette Paintjet_gpl}
\format{A vector of length 16}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Paintjet_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
