% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pastels_gpl}
\alias{Pastels_gpl}
\title{Palette Pastels_gpl}
\format{A vector of length 18}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Pastels_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
