% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reds_gpl}
\alias{Reds_gpl}
\title{Palette Reds_gpl}
\format{A vector of length 177}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Reds_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
