% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Royal_gpl}
\alias{Royal_gpl}
\title{Palette Royal_gpl}
\format{A vector of length 222}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Royal_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
