% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPalettes.R
\name{showPalette}
\alias{showPalette}
\title{Show a palette.}
\usage{
showPalette(myPal)
}
\arguments{
\item{myPal}{The palette name.}
}
\description{
Show a palette.
}
\details{
adapted from http://www.r-graph-gallery.com/42-colors-names/
}
\examples{
data(Caramel_gpl)
showPalette(myPal = Caramel_gpl)
}
