% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SVG_gpl}
\alias{SVG_gpl}
\title{Palette SVG_gpl}
\format{A vector of length 139}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
SVG_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
