% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tango_gpl}
\alias{Tango_gpl}
\title{Palette Tango_gpl}
\format{A vector of length 29}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Tango_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
