% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Topographic_gpl}
\alias{Topographic_gpl}
\title{Palette Topographic_gpl}
\format{A vector of length 191}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Topographic_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
