% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Warm_gpl}
\alias{Warm_gpl}
\title{Palette Warm_gpl}
\format{A vector of length 7}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Warm_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
