% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Web_gpl}
\alias{Web_gpl}
\title{Palette Web_gpl}
\format{A vector of length 216}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Web_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
