% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WebHex_gpl}
\alias{WebHex_gpl}
\title{Palette WebHex_gpl}
\format{A vector of length 216}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
WebHex_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
