% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WebSafe22_gpl}
\alias{WebSafe22_gpl}
\title{Palette WebSafe22_gpl}
\format{A vector of length 22}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
WebSafe22_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
