% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Windows_gpl}
\alias{Windows_gpl}
\title{Palette Windows_gpl}
\format{A vector of length 31}
\source{
\url{https://inkscape.org}

\url{https://www.gimp.org}
}
\usage{
Windows_gpl
}
\description{
A palette in hexadecimal format extracted from the Gimp or Inkscape.
}
\keyword{palette.}
